backend/requirements.txt
Django == 2.1.5
django-request-logging == 0.6.7
django_cors_headers == 2.4.0
djangorestframework == 3.9.1
djangorestframework_simplejwt == 3.3
psycopg2-binary == 2.7.7
pytz == 2018.3
redis == 3.1.0
rq == 0.13.0
